Use requirements hostname in CLI config loads
