+++ codex-rs/app-server/Cargo.toml
@@ -88,13 +88,6 @@ tracing-subscriber = { workspace = true, features = ["env-filter", "fmt", "json"
 url = { workspace = true }
 uuid = { workspace = true, features = ["serde", "v7"] }
 
-[target.'cfg(unix)'.dependencies]
-dns-lookup = { workspace = true }
-libc = { workspace = true }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-winapi-util = { workspace = true }
-
 [dev-dependencies]
 app_test_support = { workspace = true }
 base64 = { workspace = true }
