+++ codex-rs/app-server/src/config_api.rs
@@ -33,10 +33,6 @@ use codex_features::canonical_feature_for_key;
 use codex_features::feature_for_key;
 use codex_protocol::config_types::WebSearchMode;
 use codex_protocol::protocol::Op;
-#[cfg(unix)]
-use dns_lookup::AddrInfoHints;
-#[cfg(unix)]
-use dns_lookup::getaddrinfo;
 use serde_json::json;
 use std::collections::BTreeMap;
 use std::collections::BTreeSet;
@@ -45,10 +41,6 @@ use std::sync::Arc;
 use std::sync::RwLock;
 use toml::Value as TomlValue;
 use tracing::warn;
-#[cfg(windows)]
-use winapi_util::sysinfo::ComputerNameKind;
-#[cfg(windows)]
-use winapi_util::sysinfo::get_computer_name;
 
 const SUPPORTED_EXPERIMENTAL_FEATURE_ENABLEMENT: &[&str] = &[
     "apps",
@@ -116,7 +108,7 @@ impl ConfigApi {
             self.current_cli_overrides(),
             self.loader_overrides.clone(),
             self.current_cloud_requirements(),
-            app_server_requirements_hostname(),
+            codex_config::requirements_hostname(),
         )
     }
 
@@ -151,7 +143,7 @@ impl ConfigApi {
             .loader_overrides(self.loader_overrides.clone())
             .fallback_cwd(fallback_cwd)
             .cloud_requirements(self.current_cloud_requirements())
-            .requirements_hostname(app_server_requirements_hostname())
+            .requirements_hostname(codex_config::requirements_hostname())
             .build()
             .await
             .map_err(|err| JSONRPCErrorError {
@@ -371,60 +363,6 @@ pub(crate) fn apply_runtime_feature_enablement(
     }
 }
 
-pub(crate) fn app_server_requirements_hostname() -> Option<String> {
-    let kernel_hostname = gethostname::gethostname();
-    let kernel_hostname = normalize_requirements_hostname(&kernel_hostname.to_string_lossy())?;
-
-    // Remote sandbox requirements are meant to target remote hosts by DNS name,
-    // so prefer the canonical FQDN when the local resolver can provide one.
-    if let Some(fqdn) = local_fqdn_for_hostname(&kernel_hostname) {
-        return Some(fqdn);
-    }
-
-    // Some machines have only a short local hostname or resolver setup that
-    // does not return AI_CANONNAME. Keep matching behavior best-effort by
-    // falling back to the cleaned kernel hostname instead of returning None.
-    Some(kernel_hostname)
-}
-
-fn normalize_requirements_hostname(hostname: &str) -> Option<String> {
-    let hostname = hostname.trim().trim_end_matches('.');
-    (!hostname.is_empty()).then(|| hostname.to_ascii_lowercase())
-}
-
-#[cfg(unix)]
-fn local_fqdn_for_hostname(hostname: &str) -> Option<String> {
-    let hints = AddrInfoHints {
-        flags: libc::AI_CANONNAME,
-        ..AddrInfoHints::default()
-    };
-
-    getaddrinfo(Some(hostname), /*service*/ None, Some(hints))
-        .ok()?
-        .filter_map(Result::ok)
-        .filter_map(|addr| addr.canonname)
-        // getaddrinfo may return the short hostname as canonname when no FQDN
-        // is available. Treat only DNS-qualified names as an FQDN result.
-        .find_map(|hostname| normalize_fqdn_candidate(&hostname))
-}
-
-#[cfg(windows)]
-fn local_fqdn_for_hostname(_hostname: &str) -> Option<String> {
-    get_computer_name(ComputerNameKind::PhysicalDnsFullyQualified)
-        .ok()
-        .and_then(|hostname| hostname.into_string().ok())
-        .and_then(|hostname| normalize_fqdn_candidate(&hostname))
-}
-
-#[cfg(not(any(unix, windows)))]
-fn local_fqdn_for_hostname(_hostname: &str) -> Option<String> {
-    None
-}
-
-fn normalize_fqdn_candidate(hostname: &str) -> Option<String> {
-    normalize_requirements_hostname(hostname).filter(|hostname| hostname.contains('.'))
-}
-
 fn map_requirements_toml_to_api(requirements: ConfigRequirementsToml) -> ConfigRequirements {
     ConfigRequirements {
         allowed_approval_policies: requirements.allowed_approval_policies.map(|policies| {
@@ -610,27 +548,6 @@ mod tests {
         async fn reload_user_config(&self) {
             self.call_count.fetch_add(1, Ordering::Relaxed);
         }
-    }
-
-    #[test]
-    fn normalize_fqdn_candidate_accepts_dns_qualified_name() {
-        assert_eq!(
-            normalize_fqdn_candidate("runner-01.ci.example.com"),
-            Some("runner-01.ci.example.com".to_string())
-        );
-    }
-
-    #[test]
-    fn normalize_fqdn_candidate_rejects_short_name() {
-        assert_eq!(normalize_fqdn_candidate("runner-01"), None);
-    }
-
-    #[test]
-    fn normalize_fqdn_candidate_trims_trailing_dot_and_normalizes_case() {
-        assert_eq!(
-            normalize_fqdn_candidate("RUNNER-01.CI.EXAMPLE.COM."),
-            Some("runner-01.ci.example.com".to_string())
-        );
     }
 
     #[test]
