+++ codex-rs/config/Cargo.toml
@@ -19,6 +19,7 @@ codex-protocol = { workspace = true }
 codex-utils-absolute-path = { workspace = true }
 codex-utils-path = { workspace = true }
 futures = { workspace = true, features = ["alloc", "std"] }
+gethostname = { workspace = true }
 multimap = { workspace = true }
 schemars = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
@@ -31,6 +32,13 @@ toml = { workspace = true }
 toml_edit = { workspace = true }
 tracing = { workspace = true }
 wildmatch = { workspace = true }
+
+[target.'cfg(unix)'.dependencies]
+dns-lookup = { workspace = true }
+libc = { workspace = true }
+
+[target.'cfg(target_os = "windows")'.dependencies]
+winapi-util = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
