+++ codex-rs/config/src/lib.rs
@@ -14,6 +14,7 @@ pub mod permissions_toml;
 pub mod profile_toml;
 mod project_root_markers;
 mod requirements_exec_policy;
+mod requirements_hostname;
 pub mod schema;
 pub mod shell_environment;
 mod skills_config;
@@ -88,6 +89,7 @@ pub use requirements_exec_policy::RequirementsExecPolicyParseError;
 pub use requirements_exec_policy::RequirementsExecPolicyPatternTokenToml;
 pub use requirements_exec_policy::RequirementsExecPolicyPrefixRuleToml;
 pub use requirements_exec_policy::RequirementsExecPolicyToml;
+pub use requirements_hostname::requirements_hostname;
 pub use skills_config::BundledSkillsConfig;
 pub use skills_config::SkillConfig;
 pub use skills_config::SkillsConfig;
